leadmanager/requirements.txt
amqp==2.5.2
asgiref==3.2.3
astroid==2.3.3
beautifulsoup4==4.8.2
billiard==3.6.1.0
boto3==1.11.9
botocore==1.14.9
celery==4.4.0
cffi==1.13.2
cryptography==2.8
Django==3.0.2
django-rest-knox==4.1.0
djangorestframework==3.11.0
docutils==0.15.2
isort==4.3.21
jmespath==0.9.4
kombu==4.6.7
lazy-object-proxy==1.4.3
mccabe==0.6.1
psycopg2==2.8.4
pycparser==2.19
pycurl==7.43.0.4
pylint==2.4.4
python-dateutil==2.8.1
pytz==2019.3
redis==3.3.11
s3transfer==0.3.2
six==1.14.0
soupsieve==1.9.5
sqlparse==0.3.0
urllib3==1.25.8
vine==1.3.0
wrapt==1.11.2
